Pull HSV out to its own file and remove parsing.
