+++ index.js
@@ -1,3 +1,4 @@
-export {default as color, rgb, hsl, hsv} from "./src/color";
+export {default as color, rgb, hsl} from "./src/color";
 export {default as lab, hcl} from "./src/lab";
 export {default as cubehelix} from "./src/cubehelix";
+export {default as hsv} from "./src/hsv";
